+++ components/bpmn-q/modeler-component/extensions/opentosca/resources/opentosca4bpmn.json
@@ -26,6 +26,38 @@
           "type": "String"
         }
       ]
+    },
+    {
+      "name": "PrivacyPolicy",
+      "superClass": ["bpmn:Event"],
+      "properties": [
+        {
+          "name": "componentSharing",
+          "isAttr": true,
+          "type": "String"
+        },
+        {
+          "name": "cloudType",
+          "isAttr": true,
+          "type": "String"
+        }
+      ]
+    },
+    {
+      "name": "DeploymentPolicy",
+      "superClass": ["bpmn:Event"],
+      "properties": [
+        {
+          "name": "onDemand",
+          "isAttr": true,
+          "type": "String"
+        }
+      ]
+    },
+    {
+      "name": "Policy",
+      "superClass": ["bpmn:Event"],
+      "properties": []
     }
   ],
   "enumerations": [],
