+++ components/bpmn-q/modeler-component/extensions/quantme/Constants.js
@@ -29,9 +29,6 @@ export const CIRCUIT_CUTTING_SUBPROCESS = "quantme:CircuitCuttingSubprocess";
 export const CIRCUIT_CUTTING_TASK = "quantme:CircuitCuttingTask";
 export const CUTTING_RESULT_COMBINATION_TASK =
   "quantme:CuttingResultCombinationTask";
-export const POLICY = "quantme:Policy";
-export const PRIVACY_POLICY = "quantme:PrivacyPolicy";
-export const DEPLOYMENT_POLICY = "quantme:DeploymentPolicy";
 
 // Property names of the QuantME constructs
 export const ALGORITHM = "algorithm";
@@ -75,8 +72,6 @@ export const MAX_NUMBER_OF_CUTS = "maxNumberOfCuts";
 export const MAXIMUM_NUM_SUBCIRCUITS = "maxNumSubCircuits";
 
 export const ON_DEMAND = "onDemand";
-export const COMPONENT_SHARING = "componentSharing";
-export const CLOUD_TYPE = "cloudType";
 
 // endpoint paths of connected services
 export const NISQ_ANALYZER_QPU_SELECTION_PATH = "nisq-analyzer/qpu-selection";
