+++ components/bpmn-q/modeler-component/extensions/quantme/modeling/QuantMEReplaceMenuProvider.js
@@ -89,25 +89,6 @@ export default class QuantMEReplaceMenuProvider {
         );
         return Object.assign(subprocessEntries, entries);
       }
-      // add additional elements to replace policies
-      if (is(element, "bpmn:Event")) {
-        if (element.host !== undefined) {
-          if (element.host.type === "bpmn:ServiceTask") {
-            const filteredOptions = filter(
-              quantmeReplaceOptions.POLICY,
-              isDifferentType(element)
-            );
-            const policyEntries = createMenuEntries(
-              element,
-              filteredOptions,
-              self.translate,
-              self.bpmnReplace.replaceElement
-            );
-            return Object.assign(policyEntries, entries);
-          }
-        }
-      }
-
       return entries;
     };
   }
