+++ components/bpmn-q/modeler-component/extensions/quantme/modeling/QuantMEReplaceOptions.js
@@ -176,22 +176,3 @@ export var SUBPROCESS = [
     },
   },
 ];
-
-export var POLICY = [
-  {
-    label: "Deployment Policy",
-    actionName: "replace-with-deployment-policy",
-    className: "qwm bpmn-icon-deployment-policy",
-    target: {
-      type: consts.DEPLOYMENT_POLICY,
-    },
-  },
-  {
-    label: "Privacy Policy",
-    actionName: "replace-with-privacy-policy",
-    className: "qwm bpmn-icon-privacy-policy",
-    target: {
-      type: consts.PRIVACY_POLICY,
-    },
-  },
-];
