+++ components/bpmn-q/modeler-component/extensions/quantme/modeling/QuantMESVGMap.js
@@ -60,16 +60,16 @@ export function getQuantMESVG(svgId) {
     },
     PRIVACY_POLICY: {
       transform: "matrix(0.22,0,0,0.22,3,3)",
-      svg: '<svg xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape" xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd" xmlns="http://www.w3.org/2000/svg" xmlns:svg="http://www.w3.org/2000/svg" version="1.1" id="svg132" width="143.67999" height="148.16013" viewBox="0 0 143.67999 148.16012" sodipodi:docname="PrivacyPolicyPowerpoint.svg" inkscape:version="1.2.2 (732a01da63, 2022-12-09)">  <defs id="defs136">    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath148">      <path d="M 0,2.2888e-5 H 113.28 V 113.28002 H 0 Z" clip-rule="evenodd" id="path146"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath158">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path156"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath168">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path166"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath178">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path176"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath188">      <path d="M 37.2,39.36 H 82.32 V 84.48 H 37.2 Z" clip-rule="evenodd" id="path186"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath198">      <path d="M 0,0 H 23.291 V 23.286 H 0 Z" id="path196"/>    </clipPath>  </defs>  <sodipodi:namedview id="namedview134" pagecolor="#ffffff" bordercolor="#000000" borderopacity="0.25" inkscape:showpageshadow="2" inkscape:pageopacity="0.0" inkscape:pagecheckerboard="0" inkscape:deskcolor="#d1d1d1" showgrid="false" inkscape:zoom="5.6475107" inkscape:cx="70.030854" inkscape:cy="73.926376" inkscape:window-width="1920" inkscape:window-height="1111" inkscape:window-x="-9" inkscape:window-y="-9" inkscape:window-maximized="1" inkscape:current-layer="g140">    <inkscape:page x="0" y="0" id="page138" width="143.67999" height="148.16013"/>  </sodipodi:namedview>  <g id="g140" inkscape:groupmode="layer" inkscape:label="Page 1" transform="matrix(1.3333333,0,0,-1.3333333,-5.4399999,149.44)">    <g id="g152">      <g id="g154" clip-path="url(#clipPath158)">        <path d="M 11.235,111.6 C 14.922,111.6 17.91,108.61 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 H 17.91 V 8.115 C 17.91,4.4285 20.899,1.44 24.585,1.44 H 104.68 C 108.37,1.44 111.36,4.4285 111.36,8.115 111.36,11.802 108.37,14.79 104.68,14.79 H 98.01 V 104.92 C 98.01,108.61 95.021,111.6 91.335,111.6 Z M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path160"/>      </g>    </g>    <g id="g162">      <g id="g164" clip-path="url(#clipPath168)">        <path d="M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z M 17.91,104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 7.5485,111.6 4.56,108.61 4.56,104.92 4.56,101.24 7.5485,98.25 11.235,98.25 13.078,98.25 14.572,99.744 14.572,101.59 14.572,103.43 13.078,104.92 11.235,104.92 Z" style="fill:#cdcdcd;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path170"/>      </g>    </g>    <g id="g172">      <g id="g174" clip-path="url(#clipPath178)">        <path d="M 24.585,1.44 C 20.898,1.44 17.91,4.4285 17.91,8.115 V 98.25 H 11.235 C 7.5485,98.25 4.56,101.24 4.56,104.92 4.56,108.61 7.5485,111.6 11.235,111.6 H 91.335 C 95.022,111.6 98.01,108.61 98.01,104.92 V 14.79 H 104.68 C 108.37,14.79 111.36,11.802 111.36,8.115 111.36,4.4285 108.37,1.44 104.68,1.44 Z M 17.91,98.25 V 104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 M 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 M 31.26,8.115 H 24.585 C 22.742,8.115 21.247,9.6092 21.247,11.452 21.247,13.296 22.742,14.79 24.585,14.79 28.271,14.79 31.26,11.801 31.26,8.1149 31.26,4.4284 28.271,1.4399 24.585,1.4399 M 24.585,14.79 H 98.01" style="fill:none;stroke:#000000;stroke-width:0.96;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:10;stroke-dasharray:none;stroke-opacity:1" id="path180"/>      </g>    </g>    <g id="g182">      <g id="g184" clip-path="url(#clipPath188)">        <g id="g190" transform="matrix(0.62667,0,-4.7e-8,0.62667,37.2,39.36)">          <g id="g192" transform="scale(3.0913,3.092)">            <g id="g194" clip-path="url(#clipPath198)">              <path d="M 12.142,6.5359 V 5.2639 H 11.182 V 6.5359 C 10.63,6.7279 10.222,7.2559 10.222,7.9039 10.222,8.6959 10.87,9.3439 11.662,9.3439 12.454,9.3439 13.102,8.6959 13.102,7.9039 13.102,7.2799 12.694,6.7519 12.142,6.5359 Z M 8.0618,15.344 C 8.0618,17.336 9.6698,18.944 11.662,18.944 13.654,18.944 15.262,17.336 15.262,15.344 V 12.68 L 11.662,12.944 8.0618,12.68 Z M 16.702,12.584 V 15.344 C 16.702,18.128 14.446,20.384 11.662,20.384 8.8778,20.384 6.6218,18.128 6.6218,15.344 V 12.584 L 4.9418,12.464 V 3.3439 L 11.662,2.8639 18.382,3.3439 V 12.464 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path200"/>            </g>          </g>        </g>      </g>    </g>  </g></svg>'
+      svg: '<svg xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape" xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd" xmlns="http://www.w3.org/2000/svg" xmlns:svg="http://www.w3.org/2000/svg" version="1.1" id="svg132" width="143.67999" height="148.16013" viewBox="0 0 143.67999 148.16012" sodipodi:docname="PrivacyPolicyPowerpoint.svg" inkscape:version="1.2.2 (732a01da63, 2022-12-09)">  <defs id="defs136">    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath148">      <path d="M 0,2.2888e-5 H 113.28 V 113.28002 H 0 Z" clip-rule="evenodd" id="path146"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath158">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path156"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath168">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path166"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath178">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path176"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath188">      <path d="M 37.2,39.36 H 82.32 V 84.48 H 37.2 Z" clip-rule="evenodd" id="path186"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath198">      <path d="M 0,0 H 23.291 V 23.286 H 0 Z" id="path196"/>    </clipPath>  </defs>  <sodipodi:namedview id="namedview134" pagecolor="#ffffff" bordercolor="#000000" borderopacity="0.25" inkscape:showpageshadow="2" inkscape:pageopacity="0.0" inkscape:pagecheckerboard="0" inkscape:deskcolor="#d1d1d1" showgrid="false" inkscape:zoom="5.6475107" inkscape:cx="70.030854" inkscape:cy="73.926376" inkscape:window-width="1920" inkscape:window-height="1111" inkscape:window-x="-9" inkscape:window-y="-9" inkscape:window-maximized="1" inkscape:current-layer="g140">    <inkscape:page x="0" y="0" id="page138" width="143.67999" height="148.16013"/>  </sodipodi:namedview>  <g id="g140" inkscape:groupmode="layer" inkscape:label="Page 1" transform="matrix(1.3333333,0,0,-1.3333333,-5.4399999,149.44)">    <g id="g152">      <g id="g154" clip-path="url(#clipPath158)">        <path d="M 11.235,111.6 C 14.922,111.6 17.91,108.61 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 H 17.91 V 8.115 C 17.91,4.4285 20.899,1.44 24.585,1.44 H 104.68 C 108.37,1.44 111.36,4.4285 111.36,8.115 111.36,11.802 108.37,14.79 104.68,14.79 H 98.01 V 104.92 C 98.01,108.61 95.021,111.6 91.335,111.6 Z M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path160"/>      </g>    </g>    <g id="g162">      <g id="g164" clip-path="url(#clipPath168)">        <path d="M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z M 17.91,104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 7.5485,111.6 4.56,108.61 4.56,104.92 4.56,101.24 7.5485,98.25 11.235,98.25 13.078,98.25 14.572,99.744 14.572,101.59 14.572,103.43 13.078,104.92 11.235,104.92 Z" style="fill:#cdcdcd;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path170"/>      </g>    </g>    <g id="g172">      <g id="g174" clip-path="url(#clipPath178)">        <path d="M 24.585,1.44 C 20.898,1.44 17.91,4.4285 17.91,8.115 V 98.25 H 11.235 C 7.5485,98.25 4.56,101.24 4.56,104.92 4.56,108.61 7.5485,111.6 11.235,111.6 H 91.335 C 95.022,111.6 98.01,108.61 98.01,104.92 V 14.79 H 104.68 C 108.37,14.79 111.36,11.802 111.36,8.115 111.36,4.4285 108.37,1.44 104.68,1.44 Z M 17.91,98.25 V 104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 M 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 M 31.26,8.115 H 24.585 C 22.742,8.115 21.247,9.6092 21.247,11.452 21.247,13.296 22.742,14.79 24.585,14.79 28.271,14.79 31.26,11.801 31.26,8.1149 31.26,4.4284 28.271,1.4399 24.585,1.4399 M 24.585,14.79 H 98.01" style="fill:none;stroke:#000000;stroke-width:0.96;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:10;stroke-dasharray:none;stroke-opacity:1" id="path180"/>      </g>    </g>    <g id="g182">      <g id="g184" clip-path="url(#clipPath188)">        <g id="g190" transform="matrix(0.62667,0,-4.7e-8,0.62667,37.2,39.36)">          <g id="g192" transform="scale(3.0913,3.092)">            <g id="g194" clip-path="url(#clipPath198)">              <path d="M 12.142,6.5359 V 5.2639 H 11.182 V 6.5359 C 10.63,6.7279 10.222,7.2559 10.222,7.9039 10.222,8.6959 10.87,9.3439 11.662,9.3439 12.454,9.3439 13.102,8.6959 13.102,7.9039 13.102,7.2799 12.694,6.7519 12.142,6.5359 Z M 8.0618,15.344 C 8.0618,17.336 9.6698,18.944 11.662,18.944 13.654,18.944 15.262,17.336 15.262,15.344 V 12.68 L 11.662,12.944 8.0618,12.68 Z M 16.702,12.584 V 15.344 C 16.702,18.128 14.446,20.384 11.662,20.384 8.8778,20.384 6.6218,18.128 6.6218,15.344 V 12.584 L 4.9418,12.464 V 3.3439 L 11.662,2.8639 18.382,3.3439 V 12.464 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path200"/>            </g>          </g>        </g>      </g>    </g>  </g></svg>',
     },
     DEPLOYMENT_POLICY: {
       transform: "matrix(0.22,0,0,0.22,3,3)",
-      svg: '<svg xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape" xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd" xmlns="http://www.w3.org/2000/svg" xmlns:svg="http://www.w3.org/2000/svg" version="1.1" id="svg2137" width="151.03999" height="151.03999" viewBox="0 0 151.03999 151.03999" sodipodi:docname="deployment_policy.svg" inkscape:version="1.2.2 (732a01da63, 2022-12-09)">  <defs id="defs2141">    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2153">      <path d="M 0,2.2888e-5 H 113.28 V 113.28002 H 0 Z" clip-rule="evenodd" id="path2151"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2163">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2161"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2173">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2171"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2183">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2181"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2193">      <path d="M 35.28,23.52 H 84.24 V 72.48 H 35.28 Z" clip-rule="evenodd" id="path2191"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2203">      <path d="M 0,0 H 23.291 V 23.286 H 0 Z" id="path2201"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2223">      <path d="M 26.64,69.36 H 55.44 V 98.16 H 26.64 Z" clip-rule="evenodd" id="path2221"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2233">      <path d="M 0,0 H 23.291 V 23.286 H 0 Z" id="path2231"/>    </clipPath>  </defs>  <sodipodi:namedview id="namedview2139" pagecolor="#ffffff" bordercolor="#000000" borderopacity="0.25" inkscape:showpageshadow="2" inkscape:pageopacity="0.0" inkscape:pagecheckerboard="0" inkscape:deskcolor="#d1d1d1" showgrid="false" inkscape:zoom="1.5625" inkscape:cx="75.839998" inkscape:cy="75.839998" inkscape:window-width="1920" inkscape:window-height="1111" inkscape:window-x="-9" inkscape:window-y="-9" inkscape:window-maximized="1" inkscape:current-layer="g2145">    <inkscape:page x="0" y="0" id="page2143" width="151.03999" height="151.03999"/>  </sodipodi:namedview>  <g id="g2145" inkscape:groupmode="layer" inkscape:label="Page 1" transform="matrix(1.3333333,0,0,-1.3333333,0,151.04)">    <g id="g2147">      <g id="g2149" clip-path="url(#clipPath2153)">        <path d="M 0,7.629e-6 H 113.28 V 113.28001 H 0 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2155"/>      </g>    </g>    <g id="g2157">      <g id="g2159" clip-path="url(#clipPath2163)">        <path d="M 11.235,111.6 C 14.922,111.6 17.91,108.61 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 H 17.91 V 8.115 C 17.91,4.4285 20.899,1.44 24.585,1.44 H 104.68 C 108.37,1.44 111.36,4.4285 111.36,8.115 111.36,11.802 108.37,14.79 104.68,14.79 H 98.01 V 104.92 C 98.01,108.61 95.021,111.6 91.335,111.6 Z M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2165"/>      </g>    </g>    <g id="g2167">      <g id="g2169" clip-path="url(#clipPath2173)">        <path d="M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z M 17.91,104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 7.5485,111.6 4.56,108.61 4.56,104.92 4.56,101.24 7.5485,98.25 11.235,98.25 13.078,98.25 14.572,99.744 14.572,101.59 14.572,103.43 13.078,104.92 11.235,104.92 Z" style="fill:#cdcdcd;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2175"/>      </g>    </g>    <g id="g2177">      <g id="g2179" clip-path="url(#clipPath2183)">        <path d="M 24.585,1.44 C 20.898,1.44 17.91,4.4285 17.91,8.115 V 98.25 H 11.235 C 7.5485,98.25 4.56,101.24 4.56,104.92 4.56,108.61 7.5485,111.6 11.235,111.6 H 91.335 C 95.022,111.6 98.01,108.61 98.01,104.92 V 14.79 H 104.68 C 108.37,14.79 111.36,11.802 111.36,8.115 111.36,4.4285 108.37,1.44 104.68,1.44 Z M 17.91,98.25 V 104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 M 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 M 31.26,8.115 H 24.585 C 22.742,8.115 21.247,9.6092 21.247,11.452 21.247,13.296 22.742,14.79 24.585,14.79 28.271,14.79 31.26,11.801 31.26,8.1149 31.26,4.4284 28.271,1.4399 24.585,1.4399 M 24.585,14.79 H 98.01" style="fill:none;stroke:#000000;stroke-width:0.96;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:10;stroke-dasharray:none;stroke-opacity:1" id="path2185"/>      </g>    </g>    <g id="g2187">      <g id="g2189" clip-path="url(#clipPath2193)">        <g id="g2195" transform="matrix(0.68,0,-4.7e-8,0.68,35.28,23.52)">          <g id="g2197" transform="scale(3.0913,3.092)">            <g id="g2199" clip-path="url(#clipPath2203)">              <path d="M 7.3418,18.464 3.7418,16.28 11.662,11.48 15.262,13.664 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2205"/>              <path d="M 19.582,16.28 11.662,21.08 8.2538,19.016 16.174,14.216 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2207"/>              <path d="M 3.7418,14.36 V 6.6799 L 11.182,2.1679 V 10.64 L 3.7418,15.152 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2209"/>              <path d="M 14.782,9.3199 13.102,8.3599 V 10.04 L 14.782,11 Z M 12.142,10.64 V 2.1679 L 19.582,6.6799 V 15.152 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2211"/>            </g>          </g>        </g>      </g>    </g>    <path d="M 24.72,92.12 C 24.72,94.395 26.565,96.24 28.84,96.24 H 45.16 V 96.24 H 53.92 55.64 C 57.915,96.24 59.76,94.395 59.76,92.12 V 81.82 81.82 75.64 75.64 C 59.76,73.365 57.915,71.52 55.64,71.52 H 53.92 L 54.407,57.381 45.16,71.52 H 28.84 C 26.565,71.52 24.72,73.365 24.72,75.64 V 75.64 81.82 81.82 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2213"/>    <path d="M 24.72,92.12 C 24.72,94.395 26.565,96.24 28.84,96.24 H 45.16 V 96.24 H 53.92 55.64 C 57.915,96.24 59.76,94.395 59.76,92.12 V 81.82 81.82 75.64 75.64 C 59.76,73.365 57.915,71.52 55.64,71.52 H 53.92 L 54.407,57.381 45.16,71.52 H 28.84 C 26.565,71.52 24.72,73.365 24.72,75.64 V 75.64 81.82 81.82 Z" style="fill:none;stroke:#000000;stroke-width:0.96;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:10;stroke-dasharray:none;stroke-opacity:1" id="path2215"/>    <g id="g2217">      <g id="g2219" clip-path="url(#clipPath2223)">        <g id="g2225" transform="matrix(0.4,-9.4e-8,-2.4e-8,0.4,26.64,69.36)">          <g id="g2227" transform="scale(3.0913,3.092)">            <g id="g2229" clip-path="url(#clipPath2233)">              <path d="M 19.342,18.344 C 19.343,19.272 18.591,20.024 17.663,20.025 16.735,20.026 15.983,19.274 15.982,18.347 15.981,17.778 16.268,17.248 16.745,16.937 L 15.502,13.304 V 13.304 C 15.159,13.304 14.824,13.2 14.542,13.004 L 12.036,14.883 C 12.357,15.753 11.912,16.72 11.042,17.041 10.171,17.362 9.2054,16.917 8.8841,16.047 8.6159,15.32 8.879,14.505 9.521,14.072 L 7.9202,8.9839 H 7.8218 C 6.8939,8.9861 6.14,8.2357 6.1378,7.3079 6.1356,6.38 6.8859,5.6261 7.8138,5.6239 8.7416,5.6217 9.4955,6.372 9.4978,7.2999 9.499,7.8315 9.2486,8.3323 8.8226,8.6503 L 10.438,13.784 H 10.462 C 10.821,13.785 11.171,13.901 11.46,14.115 L 13.944,12.255 C 13.864,12.054 13.823,11.84 13.822,11.624 13.821,10.696 14.573,9.9436 15.501,9.9432 16.429,9.9429 17.181,10.695 17.182,11.623 17.182,12.189 16.897,12.717 16.423,13.028 L 17.662,16.664 C 18.59,16.664 19.342,17.416 19.342,18.344 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2235"/>              <path d="M 4.9418,20.024 H 3.5018 V 3.2239 H 20.062 V 4.6639 H 4.9418 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2237"/>            </g>          </g>        </g>      </g>    </g>  </g></svg>'
+      svg: '<svg xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape" xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd" xmlns="http://www.w3.org/2000/svg" xmlns:svg="http://www.w3.org/2000/svg" version="1.1" id="svg2137" width="151.03999" height="151.03999" viewBox="0 0 151.03999 151.03999" sodipodi:docname="deployment_policy.svg" inkscape:version="1.2.2 (732a01da63, 2022-12-09)">  <defs id="defs2141">    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2153">      <path d="M 0,2.2888e-5 H 113.28 V 113.28002 H 0 Z" clip-rule="evenodd" id="path2151"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2163">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2161"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2173">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2171"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2183">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2181"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2193">      <path d="M 35.28,23.52 H 84.24 V 72.48 H 35.28 Z" clip-rule="evenodd" id="path2191"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2203">      <path d="M 0,0 H 23.291 V 23.286 H 0 Z" id="path2201"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2223">      <path d="M 26.64,69.36 H 55.44 V 98.16 H 26.64 Z" clip-rule="evenodd" id="path2221"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2233">      <path d="M 0,0 H 23.291 V 23.286 H 0 Z" id="path2231"/>    </clipPath>  </defs>  <sodipodi:namedview id="namedview2139" pagecolor="#ffffff" bordercolor="#000000" borderopacity="0.25" inkscape:showpageshadow="2" inkscape:pageopacity="0.0" inkscape:pagecheckerboard="0" inkscape:deskcolor="#d1d1d1" showgrid="false" inkscape:zoom="1.5625" inkscape:cx="75.839998" inkscape:cy="75.839998" inkscape:window-width="1920" inkscape:window-height="1111" inkscape:window-x="-9" inkscape:window-y="-9" inkscape:window-maximized="1" inkscape:current-layer="g2145">    <inkscape:page x="0" y="0" id="page2143" width="151.03999" height="151.03999"/>  </sodipodi:namedview>  <g id="g2145" inkscape:groupmode="layer" inkscape:label="Page 1" transform="matrix(1.3333333,0,0,-1.3333333,0,151.04)">    <g id="g2147">      <g id="g2149" clip-path="url(#clipPath2153)">        <path d="M 0,7.629e-6 H 113.28 V 113.28001 H 0 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2155"/>      </g>    </g>    <g id="g2157">      <g id="g2159" clip-path="url(#clipPath2163)">        <path d="M 11.235,111.6 C 14.922,111.6 17.91,108.61 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 H 17.91 V 8.115 C 17.91,4.4285 20.899,1.44 24.585,1.44 H 104.68 C 108.37,1.44 111.36,4.4285 111.36,8.115 111.36,11.802 108.37,14.79 104.68,14.79 H 98.01 V 104.92 C 98.01,108.61 95.021,111.6 91.335,111.6 Z M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2165"/>      </g>    </g>    <g id="g2167">      <g id="g2169" clip-path="url(#clipPath2173)">        <path d="M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z M 17.91,104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 7.5485,111.6 4.56,108.61 4.56,104.92 4.56,101.24 7.5485,98.25 11.235,98.25 13.078,98.25 14.572,99.744 14.572,101.59 14.572,103.43 13.078,104.92 11.235,104.92 Z" style="fill:#cdcdcd;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2175"/>      </g>    </g>    <g id="g2177">      <g id="g2179" clip-path="url(#clipPath2183)">        <path d="M 24.585,1.44 C 20.898,1.44 17.91,4.4285 17.91,8.115 V 98.25 H 11.235 C 7.5485,98.25 4.56,101.24 4.56,104.92 4.56,108.61 7.5485,111.6 11.235,111.6 H 91.335 C 95.022,111.6 98.01,108.61 98.01,104.92 V 14.79 H 104.68 C 108.37,14.79 111.36,11.802 111.36,8.115 111.36,4.4285 108.37,1.44 104.68,1.44 Z M 17.91,98.25 V 104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 M 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 M 31.26,8.115 H 24.585 C 22.742,8.115 21.247,9.6092 21.247,11.452 21.247,13.296 22.742,14.79 24.585,14.79 28.271,14.79 31.26,11.801 31.26,8.1149 31.26,4.4284 28.271,1.4399 24.585,1.4399 M 24.585,14.79 H 98.01" style="fill:none;stroke:#000000;stroke-width:0.96;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:10;stroke-dasharray:none;stroke-opacity:1" id="path2185"/>      </g>    </g>    <g id="g2187">      <g id="g2189" clip-path="url(#clipPath2193)">        <g id="g2195" transform="matrix(0.68,0,-4.7e-8,0.68,35.28,23.52)">          <g id="g2197" transform="scale(3.0913,3.092)">            <g id="g2199" clip-path="url(#clipPath2203)">              <path d="M 7.3418,18.464 3.7418,16.28 11.662,11.48 15.262,13.664 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2205"/>              <path d="M 19.582,16.28 11.662,21.08 8.2538,19.016 16.174,14.216 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2207"/>              <path d="M 3.7418,14.36 V 6.6799 L 11.182,2.1679 V 10.64 L 3.7418,15.152 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2209"/>              <path d="M 14.782,9.3199 13.102,8.3599 V 10.04 L 14.782,11 Z M 12.142,10.64 V 2.1679 L 19.582,6.6799 V 15.152 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2211"/>            </g>          </g>        </g>      </g>    </g>    <path d="M 24.72,92.12 C 24.72,94.395 26.565,96.24 28.84,96.24 H 45.16 V 96.24 H 53.92 55.64 C 57.915,96.24 59.76,94.395 59.76,92.12 V 81.82 81.82 75.64 75.64 C 59.76,73.365 57.915,71.52 55.64,71.52 H 53.92 L 54.407,57.381 45.16,71.52 H 28.84 C 26.565,71.52 24.72,73.365 24.72,75.64 V 75.64 81.82 81.82 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2213"/>    <path d="M 24.72,92.12 C 24.72,94.395 26.565,96.24 28.84,96.24 H 45.16 V 96.24 H 53.92 55.64 C 57.915,96.24 59.76,94.395 59.76,92.12 V 81.82 81.82 75.64 75.64 C 59.76,73.365 57.915,71.52 55.64,71.52 H 53.92 L 54.407,57.381 45.16,71.52 H 28.84 C 26.565,71.52 24.72,73.365 24.72,75.64 V 75.64 81.82 81.82 Z" style="fill:none;stroke:#000000;stroke-width:0.96;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:10;stroke-dasharray:none;stroke-opacity:1" id="path2215"/>    <g id="g2217">      <g id="g2219" clip-path="url(#clipPath2223)">        <g id="g2225" transform="matrix(0.4,-9.4e-8,-2.4e-8,0.4,26.64,69.36)">          <g id="g2227" transform="scale(3.0913,3.092)">            <g id="g2229" clip-path="url(#clipPath2233)">              <path d="M 19.342,18.344 C 19.343,19.272 18.591,20.024 17.663,20.025 16.735,20.026 15.983,19.274 15.982,18.347 15.981,17.778 16.268,17.248 16.745,16.937 L 15.502,13.304 V 13.304 C 15.159,13.304 14.824,13.2 14.542,13.004 L 12.036,14.883 C 12.357,15.753 11.912,16.72 11.042,17.041 10.171,17.362 9.2054,16.917 8.8841,16.047 8.6159,15.32 8.879,14.505 9.521,14.072 L 7.9202,8.9839 H 7.8218 C 6.8939,8.9861 6.14,8.2357 6.1378,7.3079 6.1356,6.38 6.8859,5.6261 7.8138,5.6239 8.7416,5.6217 9.4955,6.372 9.4978,7.2999 9.499,7.8315 9.2486,8.3323 8.8226,8.6503 L 10.438,13.784 H 10.462 C 10.821,13.785 11.171,13.901 11.46,14.115 L 13.944,12.255 C 13.864,12.054 13.823,11.84 13.822,11.624 13.821,10.696 14.573,9.9436 15.501,9.9432 16.429,9.9429 17.181,10.695 17.182,11.623 17.182,12.189 16.897,12.717 16.423,13.028 L 17.662,16.664 C 18.59,16.664 19.342,17.416 19.342,18.344 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2235"/>              <path d="M 4.9418,20.024 H 3.5018 V 3.2239 H 20.062 V 4.6639 H 4.9418 Z" style="fill:#000000;fill-opacity:1;fill-rule:nonzero;stroke:none" id="path2237"/>            </g>          </g>        </g>      </g>    </g>  </g></svg>',
     },
     POLICY: {
       transform: "matrix(0.22,0,0,0.22,3,3)",
-      svg: '<svg xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape" xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd" xmlns="http://www.w3.org/2000/svg" xmlns:svg="http://www.w3.org/2000/svg" version="1.1" id="svg2773" width="151.03999" height="151.03999" viewBox="0 0 151.03999 151.03999" sodipodi:docname="policysvg.svg" inkscape:version="1.2.2 (732a01da63, 2022-12-09)">  <defs id="defs2777">    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2789">      <path d="M 0,2.2888e-5 H 113.28 V 113.28002 H 0 Z" clip-rule="evenodd" id="path2787"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2799">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2797"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2809">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2807"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2819">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2817"/>    </clipPath>  </defs>  <sodipodi:namedview id="namedview2775" pagecolor="#ffffff" bordercolor="#000000" borderopacity="0.25" inkscape:showpageshadow="2" inkscape:pageopacity="0.0" inkscape:pagecheckerboard="0" inkscape:deskcolor="#d1d1d1" showgrid="false" inkscape:zoom="1.5625" inkscape:cx="75.839998" inkscape:cy="75.839998" inkscape:window-width="1920" inkscape:window-height="1111" inkscape:window-x="-9" inkscape:window-y="-9" inkscape:window-maximized="1" inkscape:current-layer="g2781">    <inkscape:page x="0" y="0" id="page2779" width="151.03999" height="151.03999"/>  </sodipodi:namedview>  <g id="g2781" inkscape:groupmode="layer" inkscape:label="Page 1" transform="matrix(1.3333333,0,0,-1.3333333,0,151.04)">    <g id="g2783">      <g id="g2785" clip-path="url(#clipPath2789)">        <path d="M 0,7.629e-6 H 113.28 V 113.28001 H 0 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2791"/>      </g>    </g>    <g id="g2793">      <g id="g2795" clip-path="url(#clipPath2799)">        <path d="M 11.235,111.6 C 14.922,111.6 17.91,108.61 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 H 17.91 V 8.115 C 17.91,4.4285 20.899,1.44 24.585,1.44 H 104.68 C 108.37,1.44 111.36,4.4285 111.36,8.115 111.36,11.802 108.37,14.79 104.68,14.79 H 98.01 V 104.92 C 98.01,108.61 95.021,111.6 91.335,111.6 Z M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2801"/>      </g>    </g>    <g id="g2803">      <g id="g2805" clip-path="url(#clipPath2809)">        <path d="M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z M 17.91,104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 7.5485,111.6 4.56,108.61 4.56,104.92 4.56,101.24 7.5485,98.25 11.235,98.25 13.078,98.25 14.572,99.744 14.572,101.59 14.572,103.43 13.078,104.92 11.235,104.92 Z" style="fill:#cdcdcd;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2811"/>      </g>    </g>    <g id="g2813">      <g id="g2815" clip-path="url(#clipPath2819)">        <path d="M 24.585,1.44 C 20.898,1.44 17.91,4.4285 17.91,8.115 V 98.25 H 11.235 C 7.5485,98.25 4.56,101.24 4.56,104.92 4.56,108.61 7.5485,111.6 11.235,111.6 H 91.335 C 95.022,111.6 98.01,108.61 98.01,104.92 V 14.79 H 104.68 C 108.37,14.79 111.36,11.802 111.36,8.115 111.36,4.4285 108.37,1.44 104.68,1.44 Z M 17.91,98.25 V 104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 M 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 M 31.26,8.115 H 24.585 C 22.742,8.115 21.247,9.6092 21.247,11.452 21.247,13.296 22.742,14.79 24.585,14.79 28.271,14.79 31.26,11.801 31.26,8.1149 31.26,4.4284 28.271,1.4399 24.585,1.4399 M 24.585,14.79 H 98.01" style="fill:none;stroke:#000000;stroke-width:0.96;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:10;stroke-dasharray:none;stroke-opacity:1" id="path2821"/>      </g>    </g>  </g></svg>'
-    }
+      svg: '<svg xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape" xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd" xmlns="http://www.w3.org/2000/svg" xmlns:svg="http://www.w3.org/2000/svg" version="1.1" id="svg2773" width="151.03999" height="151.03999" viewBox="0 0 151.03999 151.03999" sodipodi:docname="policysvg.svg" inkscape:version="1.2.2 (732a01da63, 2022-12-09)">  <defs id="defs2777">    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2789">      <path d="M 0,2.2888e-5 H 113.28 V 113.28002 H 0 Z" clip-rule="evenodd" id="path2787"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2799">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2797"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2809">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2807"/>    </clipPath>    <clipPath clipPathUnits="userSpaceOnUse" id="clipPath2819">      <path d="M -4.77e-7,113.28 H 113.28 V -2.2888e-5 L -4.77e-7,-1.5259e-5" clip-rule="evenodd" id="path2817"/>    </clipPath>  </defs>  <sodipodi:namedview id="namedview2775" pagecolor="#ffffff" bordercolor="#000000" borderopacity="0.25" inkscape:showpageshadow="2" inkscape:pageopacity="0.0" inkscape:pagecheckerboard="0" inkscape:deskcolor="#d1d1d1" showgrid="false" inkscape:zoom="1.5625" inkscape:cx="75.839998" inkscape:cy="75.839998" inkscape:window-width="1920" inkscape:window-height="1111" inkscape:window-x="-9" inkscape:window-y="-9" inkscape:window-maximized="1" inkscape:current-layer="g2781">    <inkscape:page x="0" y="0" id="page2779" width="151.03999" height="151.03999"/>  </sodipodi:namedview>  <g id="g2781" inkscape:groupmode="layer" inkscape:label="Page 1" transform="matrix(1.3333333,0,0,-1.3333333,0,151.04)">    <g id="g2783">      <g id="g2785" clip-path="url(#clipPath2789)">        <path d="M 0,7.629e-6 H 113.28 V 113.28001 H 0 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2791"/>      </g>    </g>    <g id="g2793">      <g id="g2795" clip-path="url(#clipPath2799)">        <path d="M 11.235,111.6 C 14.922,111.6 17.91,108.61 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 H 17.91 V 8.115 C 17.91,4.4285 20.899,1.44 24.585,1.44 H 104.68 C 108.37,1.44 111.36,4.4285 111.36,8.115 111.36,11.802 108.37,14.79 104.68,14.79 H 98.01 V 104.92 C 98.01,108.61 95.021,111.6 91.335,111.6 Z M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z" style="fill:#ffffff;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2801"/>      </g>    </g>    <g id="g2803">      <g id="g2805" clip-path="url(#clipPath2809)">        <path d="M 31.26,8.115 C 31.26,11.802 28.271,14.79 24.585,14.79 22.742,14.79 21.247,13.296 21.247,11.453 21.247,9.6094 22.742,8.1151 24.585,8.1151 Z M 17.91,104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 7.5485,111.6 4.56,108.61 4.56,104.92 4.56,101.24 7.5485,98.25 11.235,98.25 13.078,98.25 14.572,99.744 14.572,101.59 14.572,103.43 13.078,104.92 11.235,104.92 Z" style="fill:#cdcdcd;fill-opacity:1;fill-rule:evenodd;stroke:none" id="path2811"/>      </g>    </g>    <g id="g2813">      <g id="g2815" clip-path="url(#clipPath2819)">        <path d="M 24.585,1.44 C 20.898,1.44 17.91,4.4285 17.91,8.115 V 98.25 H 11.235 C 7.5485,98.25 4.56,101.24 4.56,104.92 4.56,108.61 7.5485,111.6 11.235,111.6 H 91.335 C 95.022,111.6 98.01,108.61 98.01,104.92 V 14.79 H 104.68 C 108.37,14.79 111.36,11.802 111.36,8.115 111.36,4.4285 108.37,1.44 104.68,1.44 Z M 17.91,98.25 V 104.92 C 17.91,108.61 14.922,111.6 11.235,111.6 M 17.91,104.92 H 11.235 C 13.078,104.92 14.572,103.43 14.572,101.59 14.572,99.744 13.078,98.25 11.235,98.25 M 31.26,8.115 H 24.585 C 22.742,8.115 21.247,9.6092 21.247,11.452 21.247,13.296 22.742,14.79 24.585,14.79 28.271,14.79 31.26,11.801 31.26,8.1149 31.26,4.4284 28.271,1.4399 24.585,1.4399 M 24.585,14.79 H 98.01" style="fill:none;stroke:#000000;stroke-width:0.96;stroke-linecap:butt;stroke-linejoin:miter;stroke-miterlimit:10;stroke-dasharray:none;stroke-opacity:1" id="path2821"/>      </g>    </g>  </g></svg>',
+    },
   };
 
   return quantMESVGMap[svgId];
