+++ components/bpmn-q/modeler-component/extensions/quantme/modeling/index.js
@@ -16,8 +16,6 @@ import QuantMEPropertiesProvider from "./properties-provider/QuantMEPropertiesPr
 import BpmnKeyboardBinding from "./BpmnKeyboardBindings";
 import BpmnEditorActions from "./BpmnEditorActions";
 import BpmnKeyboard from "./BpmnKeyboard";
-import QuantMERules from "./QuantMERules";
-import QuantMEPaletteProvider from "./QuantMEPaletteProvider";
 
 export default {
   __init__: [
@@ -29,8 +27,6 @@ export default {
     "keyboardBindings",
     "editorActions",
     "keyboard",
-    "quantMERules",
-    "quantMEPalette",
   ],
   quantMERenderer: ["type", QuantMERenderer],
   quantMEReplaceMenu: ["type", QuantMEReplaceMenuProvider],
@@ -40,6 +36,4 @@ export default {
   keyboardBindings: ["type", BpmnKeyboardBinding],
   editorActions: ["type", BpmnEditorActions],
   keyboard: ["type", BpmnKeyboard],
-  quantMERules: ["type", QuantMERules],
-  quantMEPalette: ["type", QuantMEPaletteProvider],
 };
