+++ components/bpmn-q/modeler-component/extensions/quantme/modeling/properties-provider/QuantMEPropertiesProvider.js
@@ -15,8 +15,6 @@ import {
   VariationalQuantumAlgorithmTaskEntries,
   WarmStartingTaskEntries,
   CuttingResultCombinationTaskEntries,
-  PrivacyPolicyTaskEntries,
-  DeploymentPolicyTaskEntries,
 } from "./QuantMETaskProperties";
 import * as configConsts from "../../../../editor/configurations/Constants";
 import { instance as dataObjectConfigs } from "../../configurations/DataObjectConfigurations";
@@ -149,12 +147,6 @@ function QuantMEProps(element) {
       return WarmStartingTaskEntries(element);
     case consts.CUTTING_RESULT_COMBINATION_TASK:
       return CuttingResultCombinationTaskEntries(element);
-    case consts.PRIVACY_POLICY:
-      return PrivacyPolicyTaskEntries(element);
-    case consts.DEPLOYMENT_POLICY:
-      return DeploymentPolicyTaskEntries(element);
-    case consts.POLICY:
-      return;
     default:
       console.log("Unsupported QuantME element of type: ", element.type);
   }
