+++ components/bpmn-q/modeler-component/extensions/quantme/modeling/properties-provider/QuantMEPropertyEntries.js
@@ -11,7 +11,7 @@
 
 import React from "@bpmn-io/properties-panel/preact/compat";
 
-import { TextFieldEntry, SelectEntry, CheckboxEntry } from "@bpmn-io/properties-panel";
+import { TextFieldEntry, SelectEntry } from "@bpmn-io/properties-panel";
 import * as consts from "../../Constants";
 import { useService } from "bpmn-js-properties-panel";
 import { HiddenTextFieldEntry } from "../../../../editor/popup/HiddenTextFieldEntry";
@@ -1468,119 +1468,3 @@ export function OptimizationEntry({ element }) {
     />
   );
 }
-
-export function CloudTypeEntry({ element }) {
-  const modeling = useService("modeling");
-  const translate =
-    useService("translate") ||
-    function (str) {
-      return str;
-    };
-  const debounce = useService("debounceInput");
-
-  const getValue = function () {
-    return element.businessObject.cloudType;
-  };
-
-  const setValue = function (newValue) {
-    return modeling.updateProperties(element, {
-      cloudType: newValue,
-    });
-  };
-
-  const selectOptions = [
-    { value: "public", label: "Public" },
-    { value: "private", label: "Private" },
-  ];
-
-  const getOptions = function () {
-    return selectOptions;
-  };
-
-  return (
-    <SelectEntry
-      id={consts.CLOUD_TYPE}
-      label={translate("Cloud Type")}
-      getValue={getValue}
-      setValue={setValue}
-      getOptions={getOptions}
-      debounce={debounce}
-    />
-  );
-}
-
-export function OnDemandEntry({ element }) {
-  const modeling = useService("modeling");
-  const translate =
-    useService("translate") ||
-    function (str) {
-      return str;
-    };
-  const debounce = useService("debounceInput");
-
-  const getValue = function () {
-    return element.businessObject.onDemand;
-  };
-
-  const setValue = function (newValue) {
-    let onDemandHostAttribute = element.host.businessObject.onDemand;
-    if (onDemandHostAttribute !== undefined && onDemandHostAttribute !== newValue) {
-      element.host.businessObject.onDemand = newValue;
-    }
-    return modeling.updateProperties(element, {
-      onDemand: newValue,
-    });
-  };
-
-  return (
-    <CheckboxEntry
-      id={consts.ON_DEMAND}
-      label={translate("Deploy on-demand")}
-      getValue={getValue}
-      setValue={setValue}
-      debounce={debounce}
-    />
-  );
-}
-
-export function ComponentSharingEntry({ element }) {
-  const modeling = useService("modeling");
-  const translate =
-    useService("translate") ||
-    function (str) {
-      return str;
-    };
-  const debounce = useService("debounceInput");
-
-  const getValue = function () {
-    return element.businessObject.componentSharing;
-  };
-
-  const setValue = function (newValue) {
-    return modeling.updateProperties(element, {
-      componentSharing: newValue,
-    });
-  };
-
-  const selectOptions = [
-    { value: "dedicated", label: "Dedicated Component" },
-    { value: "shared", label: "Shared Component" },
-    { value: "tenantIsolated", label: "Tenant-isolated Component" },
-  ];
-
-  const getOptions = function () {
-    return selectOptions;
-  };
-
-  return (
-    <SelectEntry
-      id={consts.COMPONENT_SHARING}
-      label={translate("Component Sharing")}
-      getValue={getValue}
-      setValue={setValue}
-      getOptions={getOptions}
-      debounce={debounce}
-    />
-  );
-}
-
