+++ components/bpmn-q/modeler-component/extensions/quantme/resources/quantum4bpmn.json
@@ -585,38 +585,6 @@
           "type": "String"
         }
       ]
-    },
-    {
-      "name": "PrivacyPolicy",
-      "superClass": [ "bpmn:Event" ],
-      "properties": [
-        {
-          "name": "componentSharing",
-          "isAttr": true,
-          "type": "String"
-        },
-        {
-          "name": "cloudType",
-          "isAttr": true,
-          "type": "String"
-        }
-      ]
-    },
-    {
-      "name": "DeploymentPolicy",
-      "superClass": [ "bpmn:Event" ],
-      "properties": [
-        {
-          "name": "onDemand",
-          "isAttr": true,
-          "type": "String"
-        }
-      ]
-    },
-    {
-      "name": "Policy",
-      "superClass": [ "bpmn:Event" ],
-      "properties": []
     }
   ],
   "enumerations": [],
