+++ components/bpmn-q/webpack.config.js
